chore: refactor DigestHelper class
