+++ src/main/java/com/limechain/sync/fullsync/FullSyncMachine.java
@@ -94,8 +94,8 @@ public class FullSyncMachine {
         blockState.storeRuntime(lastFinalizedBlockHash, runtime);
 
         int startNumber = syncState.getLastFinalizedBlockNumber()
-            .add(BigInteger.ONE)
-            .intValue();
+                .add(BigInteger.ONE)
+                .intValue();
 
         int blocksToFetch = 100;
         List<Block> receivedBlocks = requestBlocks(startNumber, blocksToFetch);
@@ -103,7 +103,7 @@ public class FullSyncMachine {
         while (!receivedBlocks.isEmpty()) {
             executeBlocks(receivedBlocks, trieAccessor);
             log.info("Executed blocks from " + receivedBlocks.getFirst().getHeader().getBlockNumber()
-                + " to " + receivedBlocks.getLast().getHeader().getBlockNumber());
+                    + " to " + receivedBlocks.getLast().getHeader().getBlockNumber());
             startNumber += blocksToFetch;
             receivedBlocks = requestBlocks(startNumber, blocksToFetch);
         }
@@ -169,18 +169,18 @@ public class FullSyncMachine {
             final int BODY = 0b0000_0010;
             final int JUSTIFICATION = 0b0001_0000;
             SyncMessage.BlockResponse response = networkService.makeBlockRequest(new BlockRequestDto(
-                HEADER | BODY | JUSTIFICATION,
-                null, // no hash, number instead
-                start,
-                SyncMessage.Direction.Ascending,
-                amount
+                    HEADER | BODY | JUSTIFICATION,
+                    null, // no hash, number instead
+                    start,
+                    SyncMessage.Direction.Ascending,
+                    amount
             ));
 
             List<SyncMessage.BlockData> blockDatas = response.getBlocksList();
 
             return blockDatas.stream()
-                .map(FullSyncMachine::protobufDecodeBlock)
-                .toList();
+                    .map(FullSyncMachine::protobufDecodeBlock)
+                    .toList();
         } catch (Exception ex) {
             log.info("Error while fetching blocks, trying to fetch again");
             if (!this.networkService.updateCurrentSelectedPeerWithNextBootnode()) {
@@ -198,9 +198,9 @@ public class FullSyncMachine {
 
         // Protobuf decode the block body
         List<Extrinsics> extrinsicsList = blockData.getBodyList().stream()
-            .map(bs -> ScaleUtils.Decode.decode(bs.toByteArray(), ScaleCodecReader::readByteArray))
-            .map(Extrinsics::new)
-            .toList();
+                .map(bs -> ScaleUtils.Decode.decode(bs.toByteArray(), ScaleCodecReader::readByteArray))
+                .map(Extrinsics::new)
+                .toList();
 
         BlockBody blockBody = new BlockBody(extrinsicsList);
 
@@ -250,8 +250,8 @@ public class FullSyncMachine {
             }
 
             boolean blockUpdatedRuntime = Arrays.stream(blockHeader.getDigest())
-                .map(HeaderDigest::getType)
-                .anyMatch(type -> type.equals(DigestType.RUN_ENV_UPDATED));
+                    .map(HeaderDigest::getType)
+                    .anyMatch(type -> type.equals(DigestType.RUN_ENV_UPDATED));
 
             if (blockUpdatedRuntime) {
                 log.info("Runtime updated, updating the runtime code");
@@ -264,9 +264,9 @@ public class FullSyncMachine {
 
     private Runtime buildRuntimeFromState(TrieAccessor trieAccessor) {
         return trieAccessor
-            .findStorageValue(Nibbles.fromBytes(":code".getBytes()))
-            .map(wasm -> runtimeBuilder.buildRuntime(wasm, trieAccessor))
-            .orElseThrow(() -> new RuntimeException("Runtime code not found in the trie"));
+                .findStorageValue(Nibbles.fromBytes(":code".getBytes()))
+                .map(wasm -> runtimeBuilder.buildRuntime(wasm, trieAccessor))
+                .orElseThrow(() -> new RuntimeException("Runtime code not found in the trie"));
     }
 
     private boolean checkInherents(Block block) {
@@ -286,13 +286,13 @@ public class FullSyncMachine {
      */
     private static boolean isBlockGoodToExecute(byte[] checkInherentsOutput) {
         var data = ScaleUtils.Decode.decode(
-            ArrayUtils.subarray(checkInherentsOutput, 2, checkInherentsOutput.length),
-            new ListReader<>(
-                new PairReader<>(
-                    scr -> new String(scr.readByteArray(8)),
-                    scr -> new String(scr.readByteArray())
+                ArrayUtils.subarray(checkInherentsOutput, 2, checkInherentsOutput.length),
+                new ListReader<>(
+                        new PairReader<>(
+                                scr -> new String(scr.readByteArray(8)),
+                                scr -> new String(scr.readByteArray())
+                        )
                 )
-            )
         );
 
         boolean goodToExecute;
